refactor(configuration): moving shared constants to packages/api
