+++ packages/main/src/plugin/configuration-registry.ts
@@ -21,11 +21,11 @@ import * as path from 'node:path';
 
 import type * as containerDesktopAPI from '@podman-desktop/api';
 
+import { CONFIGURATION_DEFAULT_SCOPE } from '/@api/configuration/constants.js';
 import type { NotificationCardOptions } from '/@api/notification.js';
 
 import type { ApiSenderType } from './api.js';
 import { ConfigurationImpl } from './configuration-impl.js';
-import { CONFIGURATION_DEFAULT_SCOPE } from './configuration-registry-constants.js';
 import type { Directories } from './directories.js';
 import type { Event } from './events/emitter.js';
 import { Emitter } from './events/emitter.js';
