+++ packages/renderer/src/lib/onboarding/OnboardingItem.svelte
@@ -4,10 +4,10 @@ import type { Unsubscriber } from 'svelte/store';
 
 import { configurationProperties } from '/@/stores/configurationProperties';
 import { context } from '/@/stores/context';
+import { CONFIGURATION_ONBOARDING_SCOPE } from '/@api/configuration/constants.js';
 import type { OnboardingStepItem } from '/@api/onboarding';
 
 import type { IConfigurationPropertyRecordedSchema } from '../../../../main/src/plugin/configuration-registry';
-import { CONFIGURATION_ONBOARDING_SCOPE } from '../../../../main/src/plugin/configuration-registry-constants';
 import type { ContextUI } from '../context/context';
 import Markdown from '../markdown/Markdown.svelte';
 import PreferencesRenderingItem from '../preferences/PreferencesRenderingItem.svelte';
